admin and pedido-pagamento

+++ src/app/controllers/productcontroller.js
@@ -4,7 +4,8 @@ const mysqConnection = require("../../database");
 const jwt = require('jsonwebtoken');
 
 const { verifyToken } = require("../middleware/jwtmiddleware")
-const { insertproduto,checkNumserieExists, updatePatrimony, deletePatrimony } = require("../models/productModel");
+const { insertproduto,checkNumserieExists, updateProduto, deletePatrimony } = require("../models/productModel");
+const { isAdmin } = require("../models/adminModel");
 const { compareSync } = require("bcrypt");
 
 const router = express.Router();
@@ -12,35 +13,17 @@ const router = express.Router();
 // Rota para cadastrar um patrimônio
 router.post('/register',verifyToken, async (req, res) => {
     try {
-        const { name, valor, custo, quantidade, descricao, categoria} = req.body;
+        const user_id = req.userId;
+
+        const userIsAdmin = await isAdmin(user_id);
+
+        if(!userIsAdmin){
+            return res.status(403).json({message: "Você nao tem permissão para realizar esta ação"})
+        }
+
+        const { name, valorcpf, valorcnpj, custo, quantidade, descricao, categoria} = req.body;
         
-        insertproduto(name, valor, custo, quantidade, descricao, categoria, res);
-    } catch (error) {
-        console.error(error);
-        res.status(500).json({ message: "Erro interno do servidor." });
-    }
-});
-
-// Rota para obter patrimônios de um usuário específico
-router.get('/list', async (req, res) => {
-    try {
-        const  user_id  = req.userId;
-
-        // Consultar o banco de dados para obter os patrimônios do usuário
-        const getproductsQuery = "SELECT p.prod_name, p.prod_valorCPF, p.prod_quantidade, p.prod_descricao, c.cat_name, c.cat_descricao FROM tbl_produto p JOIN tbl_categoria c ON p.cat_id = c.cat_id";
-
-        mysqConnection.query(getproductsQuery, [user_id], (err, results) => {
-            if (err) {
-                console.error(err);
-                return res.status(400).json({ message: "Erro ao buscar os produtos no banco de dados." });
-            }
-
-            if (results.length > 0) {
-                res.status(200).json({ produtos: results });
-            } else {
-                res.status(404).json({ message: "Nenhum patrimônio encontrado para o usuário especificado." });
-            }
-        });
+        insertproduto(name,  valorcpf, valorcnpj, custo, quantidade, descricao, categoria, res);
     } catch (error) {
         console.error(error);
         res.status(500).json({ message: "Erro interno do servidor." });
@@ -50,7 +33,7 @@ router.get('/list', async (req, res) => {
 router.get('/list', verifyToken, async (req, res) => {
     try {
         
-        const user_type = req.usertype;
+        const user_type = req.userType;
 
 
         // Construa a consulta SQL com base no tipo de usuário
@@ -83,48 +66,21 @@ router.get('/list', verifyToken, async (req, res) => {
 
 
 // Rota para atualizar um patrimônio
-router.put('/update/:patr_id', verifyToken, async (req, res) => {
+router.put('/update/:produto_id', verifyToken, async (req, res) => {
     try {
-        const patr_id = req.params.patr_id;
-        const user_id  = req.userId;
+        const produto_id = req.params.produto_id; // Obtém o ID do produto da URL
+
+
+        const userIsAdmin = await isAdmin(user_id);
+
+        if (!userIsAdmin) {
+            // Se o usuário não for um administrador, retorne um erro
+            return res.status(403).json({ message: "Você não tem permissão para realizar esta ação. Apenas administradores podem registrar produtos." });
+        }
+
+        const { name, valor, custo, quantidade, descricao, categoria } = req.body;
         
-        const {
-            valor,
-            tamanho,
-            quantidade,
-            dt_compra,
-            passativ,
-            garantia,
-            numserie,
-            notafiscal,
-            local,
-            imei,
-            ativoinativo,
-            descricao,
-            responsavel,
-            documentacao,
-            depreciacao
-        } = req.body;
-
-        const updateValues = {
-            patr_valor: valor,
-            patr_tamanho: tamanho,
-            patr_quantidade: quantidade,
-            patr_dt_compra: dt_compra,
-            patr_passativ: passativ,
-            patr_garantia: garantia,
-            patr_numserie: numserie,
-            patr_notafiscal: notafiscal,
-            patr_local: local,
-            patr_imei: imei,
-            patr_ativoinativo: ativoinativo,
-            patr_descricao: descricao,
-            patr_responsavel: responsavel,
-            patr_documentacao: documentacao,
-            patr_depreciacao: depreciacao
-        };
-
-        updatePatrimony(updateValues, patr_id, user_id)
+        updateProduto(produto_id, name, valor, custo, quantidade, descricao, categoria, res);
     } catch (error) {
         console.error(error);
         res.status(500).json({ message: "Internal Server Error: " + error.message });
@@ -132,10 +88,17 @@ router.put('/update/:patr_id', verifyToken, async (req, res) => {
 });
 
 // Rota para deletar um patrimônio
-router.delete('/delete/:patrid', verifyToken, async (req, res) => {
+router.delete('/delete/:prodid', verifyToken, async (req, res) => {
     try {
-        const patr_id = req.params.patrid;
+        const patr_id = req.params.prodid;
         const  user_id  = req.userId;
+
+        const userIsAdmin = await isAdmin(user_id);
+
+        if (!userIsAdmin) {
+            // Se o usuário não for um administrador, retorne um erro
+            return res.status(403).json({ message: "Você não tem permissão para realizar esta ação. Apenas administradores podem registrar produtos." });
+        }
 
         // Deletar o patrimônio no banco de dados
         deletePatrimony(patr_id, user_id, res);
